+++ app/models/dynamic/migration_handler.rb
@@ -135,7 +135,7 @@ module Dynamic
 
       mode = 'update'
       gs = migration_generator.generator_script(self.class, mode)
-      fn = migration_generator.write_db_migration gs, table_name, migration_generator.migration_version, mode: mode
+      fn = migration_generator.write_db_migration(gs, table_name, migration_generator.migration_version, mode:)
       @do_migration = fn
     end
 
@@ -150,7 +150,7 @@ module Dynamic
       mg.app_type_name = app_type_name
       mode = 'create_or_update'
       gs = mg.generator_script(self.class, mode)
-      mg.write_db_migration(gs, table_name, mg.migration_version, mode: mode, export_type: export_type)
+      mg.write_db_migration(gs, table_name, mg.migration_version, mode:, export_type:)
     end
 
     #
@@ -211,11 +211,11 @@ module Dynamic
       @migration_generator =
         Admin::MigrationGenerator.new(
           db_migration_schema,
-          table_name: table_name,
+          table_name:,
           class_name: full_implementation_class_name,
           dynamic_def: self,
           all_implementation_fields: all_implementation_fields(ignore_errors: false),
-          table_comments: table_comments,
+          table_comments:,
           no_master_association: implementation_no_master_association,
           prev_table_name: table_name_before_last_save,
           belongs_to_model: btm,
@@ -224,7 +224,7 @@ module Dynamic
           view_sql_changed: view_sql_changed?,
           all_referenced_tables: art,
           resource_type: self.class.name.underscore.to_sym,
-          allow_migrations: allow_migrations
+          allow_migrations:
         )
     end
 
@@ -235,7 +235,11 @@ module Dynamic
     def init_schema_name
       return if disabled?
 
-      self.schema_name = db_migration_schema
+      self.schema_name = if persisted?
+                           schema_name_in_db
+                         else
+                           db_migration_schema
+                         end
     end
   end
 end