+++ app/models/external_identifier.rb
@@ -252,7 +252,7 @@ class ExternalIdentifier < ActiveRecord::Base
   def update_tracker_events
     return unless label && !disabled
 
-    Tracker.add_record_update_entries name.singularize, current_admin, 'record'
+    Tracker.add_record_update_entries tracker_name, current_admin, 'record'
     Classification::Protocol.reset_memos
     # flag items are added when item flag names are added to the list
     # Tracker.add_record_update_entries self.name.singularize, current_admin, 'flag'