+++ app/models/save_triggers/save_triggers_base.rb
@@ -2,7 +2,7 @@
 
 # Base class for save triggers
 class SaveTriggers::SaveTriggersBase
-  attr_accessor :config, :user, :item, :master, :model_defs, :this_config
+  attr_accessor :config, :user, :item, :master, :model_defs, :this_config, :in_master
 
   def initialize(config, item)
     self.config = config
@@ -48,7 +48,12 @@ class SaveTriggers::SaveTriggersBase
       raise FphsException, 'with_result.from returns no result - check return: return_result has been specified'
     end
 
-    with_result[:attributes].each do |to, from|
+    with_attrs = with_result[:attributes].dup
+
+    create_with_ei = with_attrs.delete(:embedded_item) if with_attrs[:embedded_item]
+    vals[:embedded_item] ||= {} if create_with_ei
+
+    with_attrs.each do |to, from|
       sval = if from.include? '{{'
                FieldDefaults.calculate_default(source, from)
              else
@@ -56,6 +61,39 @@ class SaveTriggers::SaveTriggersBase
              end
 
       vals[to] = sval
+    end
+
+    create_with_ei&.each do |to, from|
+      sval = if from.include? '{{'
+               FieldDefaults.calculate_default(source, from)
+             else
+               source.attributes[from.to_s]
+             end
+
+      vals[:embedded_item][to] = sval
+    end
+  end
+
+  def handle_with_attributes(create_with, vals)
+    return unless create_with
+
+    create_with = create_with.dup
+    create_with_ei = create_with.delete(:embedded_item) if create_with[:embedded_item]
+    vals[:embedded_item] ||= {} if create_with_ei
+
+    create_with.each do |fn, def_val|
+      res = FieldDefaults.calculate_default @item, def_val
+      vals[fn] = res
+
+      if fn.to_sym == :master_id
+        self.in_master = Master.find(res)
+        in_master.current_user = @item.current_user
+      end
+    end
+
+    create_with_ei&.each do |fn, def_val|
+      res = FieldDefaults.calculate_default @item, def_val
+      vals[:embedded_item][fn] = res
     end
   end